refactor: major refactor of package layout
