+++ app/src/main/java/com/goldenraven/devkitwallet/DevkitWalletActivity.kt
@@ -11,8 +11,8 @@ import android.util.Log
 import androidx.activity.compose.setContent
 import com.goldenraven.devkitwallet.data.Repository
 import com.goldenraven.devkitwallet.data.Wallet
-import com.goldenraven.devkitwallet.ui.HomeNavigation
-import com.goldenraven.devkitwallet.ui.intro.CreateWalletNavigation
+import com.goldenraven.devkitwallet.navigation.HomeNavigation
+import com.goldenraven.devkitwallet.ui.screens.intro.CreateWalletNavigation
 
 private const val TAG = "DevkitWalletActivity"
