+++ app/src/main/java/com/goldenraven/devkitwallet/navigation/HomeNavigation.kt
@@ -3,13 +3,18 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
  */
 
-package com.goldenraven.devkitwallet.ui
+package com.goldenraven.devkitwallet.navigation
 
 import androidx.compose.animation.AnimatedContentScope
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.animation.core.tween
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
+import com.goldenraven.devkitwallet.ui.Screen
+import com.goldenraven.devkitwallet.ui.screens.WalletRoot
+import com.goldenraven.devkitwallet.ui.screens.drawer.AboutScreen
+import com.goldenraven.devkitwallet.ui.screens.drawer.ElectrumScreen
+import com.goldenraven.devkitwallet.ui.screens.drawer.RecoveryPhraseScreen
 import com.google.accompanist.navigation.animation.AnimatedNavHost
 import com.google.accompanist.navigation.animation.composable
 import com.google.accompanist.navigation.animation.rememberAnimatedNavController
