+++ app/src/main/java/com/goldenraven/devkitwallet/ui/screens/intro/CreateWalletNavigation.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
  */
 
-package com.goldenraven.devkitwallet.ui.intro
+package com.goldenraven.devkitwallet.ui.screens.intro
 
 import androidx.compose.animation.AnimatedContentScope
 import androidx.compose.animation.ExperimentalAnimationApi
