--- app/src/main/java/com/goldenraven/devkitwallet/ui/wallet/WalletNavigation.kt
@@ -1,129 +1,0 @@
-/*
- * Copyright 2020-2022 thunderbiscuit and contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
- */
-
-package com.goldenraven.devkitwallet.ui.wallet
-
-import androidx.compose.animation.AnimatedContentScope
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavHostController
-import com.goldenraven.devkitwallet.ui.Screen
-import com.goldenraven.devkitwallet.ui.wallet.*
-import com.google.accompanist.navigation.animation.AnimatedNavHost
-import com.google.accompanist.navigation.animation.composable
-import com.google.accompanist.navigation.animation.rememberAnimatedNavController
-
-@OptIn(ExperimentalAnimationApi::class)
-@Composable
-fun WalletNavigation(paddingValues: PaddingValues) {
-    val navController: NavHostController = rememberAnimatedNavController()
-    val animationDuration = 400
-
-    AnimatedNavHost(
-        navController = navController,
-        startDestination = Screen.HomeScreen.route,
-    ) {
-
-        composable(
-            route = Screen.HomeScreen.route,
-            exitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            popEnterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            },
-        ) { HomeScreen(navController, paddingValues) }
-
-        composable(
-            route = Screen.ReceiveScreen.route,
-            enterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            exitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            },
-            popEnterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            popExitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            }
-        ) { ReceiveScreen(navController, paddingValues) }
-
-        composable(
-            route = Screen.SendScreen.route,
-            enterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            exitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            },
-            popEnterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            popExitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            }
-        ) { SendScreen(navController, paddingValues) }
-
-        composable(
-            route = "${Screen.RBFScreen.route}/txid={txid}",
-            enterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            exitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            },
-            popEnterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            popExitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            }
-        ) { backStackEntry ->
-            backStackEntry.arguments?.getString("txid")?.let {
-                RBFScreen(navController, paddingValues, backStackEntry.arguments?.getString("txid"))
-            }
-        }
-
-        composable(
-            route = Screen.TransactionsScreen.route,
-            enterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            exitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            },
-            popEnterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            popExitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            }
-        ) { TransactionsScreen(navController, paddingValues) }
-
-        composable(
-            route = "${Screen.TransactionScreen.route}/txid={txid}",
-            enterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            exitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            },
-            popEnterTransition = {
-                slideIntoContainer(AnimatedContentScope.SlideDirection.Start, animationSpec = tween(animationDuration))
-            },
-            popExitTransition = {
-                slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
-            }
-        ) { backStackEntry ->
-            backStackEntry.arguments?.getString("txid")?.let {
-                TransactionScreen(navController, paddingValues, backStackEntry.arguments?.getString("txid"))
-            }
-        }
-    }
-}
