+++ app/src/main/java/com/goldenraven/devkitwallet/utils/FormatInBtc.kt
@@ -1,4 +1,4 @@
-package com.goldenraven.devkitwallet.utilities
+package com.goldenraven.devkitwallet.utils
 
 import java.text.DecimalFormat
