+++ app/src/main/java/com/goldenraven/devkitwallet/utils/SharedPreferencesManager.kt
@@ -1,5 +1,5 @@
 
-package com.goldenraven.devkitwallet.utilities
+package com.goldenraven.devkitwallet.utils
 
 import android.content.SharedPreferences
